Done.
I don't think that this will make it so that I can work on this project
remotely, as Stencyl  and Github haven't played along nicely in the
past. Here it is anyway. Commit. I need to animate the nests if I want
to do a patch later, anyway. Just a heads up to future me.